Update style.css
Re-created the CSS after receiving some feedback. Changes h2 to h1, fixed the container so it was probably centered and had a fixed width. The fixed with then made it so no matter the page size, the container stayed the same. Once I had a fixed container it was easier to adjust the elements inside of it.